refactor: 맞는 관심사로 메서드 이동
