+++ src/main/java/com/example/testysavingsbe/domain/recipe/service/usecase/RecipeQueryUseCase.java
@@ -1,17 +1,23 @@
 package com.example.testysavingsbe.domain.recipe.service.usecase;
 
+import com.example.testysavingsbe.domain.recipe.dto.response.AIChangeRecipeResponse;
 import com.example.testysavingsbe.domain.recipe.entity.CustomRecipe;
 import com.example.testysavingsbe.domain.recipe.entity.Recipe;
 import com.example.testysavingsbe.domain.user.entity.User;
+import lombok.Builder;
 import org.springframework.data.domain.Page;
 
 import java.util.List;
 
 public interface RecipeQueryUseCase {
+
     CustomRecipe getCustomRecipe(User user, String id);
 
+    AIChangeRecipeResponse createRecipeFromIngredients(User user,
+        RecipeFromIngredientsRequest request);
 
-    // todo 추천 레시피 받기
+    AIChangeRecipeResponse simplifyRecipe(User user, String recipeId);
+
     Page<Recipe> getRecommendedRecipe(User user, int page, int size);
 
     Recipe getRecipeById(User user, String id);
@@ -21,4 +27,14 @@ public interface RecipeQueryUseCase {
     boolean checkBookmarked(User user, String recipeId);
 
     List<Recipe> getBookmarkedRecipes(User user);
+
+    @Builder
+    record RecipeFromIngredientsRequest(
+        String originalRecipeId,
+        List<String> dislikeIngredients,
+        List<String> basicSeasoning,
+        List<String> mustUseIngredients
+    ) {
+
+    }
 }
